Add initial set of tests for DefinitionPrinter
We piggy-back on the partial schema parsing to reduce the work required to write the tests. Basically we expect a printed partial schema to be equal to the input source. Some issues with these tests *may* be due to the fact that we're parsing incorrectly, but I think it's more likely that any problems lie in the DefinitionPrinter itself.

Closes #211